rename: 분자로 사이드바 모달 이동
